bugfixes; more to come
