+++ ebagis/models/upload.py
@@ -12,9 +12,11 @@ from drf_chunked_upload.models import ChunkedUpload
 class Upload(ChunkedUpload):
     content_type = models.ForeignKey(ContentType)
     object_id = models.UUIDField(null=True, blank=True)
-    content_object = GenericForeignKey('content_type', 'object_id', for_concrete_model=False)
+    content_object = GenericForeignKey('content_type', 'object_id',
+                                       for_concrete_model=False)
     is_update = models.BooleanField(default=False)
-    parent_object_content_type = models.ForeignKey(ContentType)
+    parent_object_content_type = \
+        models.ForeignKey(ContentType, related_name="upload_parent")
     parent_object_id = models.UUIDField(null=True, blank=True)
     comment = models.TextField(blank=True)
     task = models.ForeignKey(TaskMeta, related_name='aoi_upload',
